+++ public-api/examples/list_public_api.rs
@@ -1,14 +1,12 @@
 use std::error::Error;
 
-use public_api::{Options, PublicApi};
-
 fn main() -> Result<(), Box<dyn Error>> {
     let json_path = rustdoc_json::Builder::default()
         .toolchain("nightly")
         .manifest_path("test-apis/example_api-v0.2.0/Cargo.toml")
         .build()?;
 
-    let public_api = PublicApi::from_rustdoc_json(json_path, Options::default())?;
+    let public_api = public_api::Builder::from_rustdoc_json(json_path).build()?;
 
     for public_item in public_api.items() {
         println!("{public_item}");